fix(frontend+api): budget caps in minor units, dashboard math parity, UX money fixes
Budget caps unit standardization (the 100x class):
- category_budgets.amount and household_budgets.budgeted now store integer
  MINOR units like every transaction-derived figure (migration 052 converts
  existing 2-decimal-currency rows). Before, caps were stored as typed
  (major) while spend was minor: backend risk stats compared a 200 USD cap
  against 12050 cents, and the frontend rendered the same fields with two
  contradictory ÷100 conventions — one always 100× off. CLP masked it.
- Caps editor converts at the input boundary (storedToMajor/majorToStored);
  BudgetBars/risk band/editor hints now agree.
- BudgetBars compares caps only against spending in the SAME currency.
- Unified the category-budgets query key — cap edits now refresh the
  dashboard bars.

Dashboard/client math parity:
- Client totals mirror the backend counts-toward-totals rule (transfer-TYPED
  exclusion, refund pairs, reimbursement groups, orphans).
- Trip money mutations invalidate the trips-list prefix (stale
  your_net_balance).
- Reimbursement dialog: one-minor-unit tolerance, same-currency-only sums,
  app locale dates.
- Trip dialogs: double-tap guard (duplicate expenses), zero-decimal steps,
  a11y group labels; transfer aria/parens fixed; currency-aware net-zero
  chip; sameAsHogar no longer collapses on all-zero months.

Backend hardening:
- users cache blob round-trips feature_trips_enabled (intermittent trips
  403s on cache hits).
- Transaction edits authorize owner-or-shared (partner personal rows are
  untouchable).
- WhatsApp PIN: CSPRNG + 3/hour rate limit; plaid router error hygiene;
  server-side currency defaults; store_credentials upsert;
  Mapped[Decimal] on amount; merchant-training warn-logs.
- test_patch_me migrated from rotted AsyncMock to real DB.

+++ backend/modules/transactions/models.py
@@ -1,4 +1,5 @@
 import uuid
+from decimal import Decimal
 from datetime import datetime
 from sqlalchemy import Boolean, DateTime, ForeignKey, Integer, JSON, Numeric, String, Text, func
 from sqlalchemy.dialects.postgresql import JSONB, UUID
@@ -17,7 +18,10 @@ class Transaction(Base):
     )
     merchant_id: Mapped[uuid.UUID | None] = mapped_column(ForeignKey("merchants.id"), nullable=True)
     raw_merchant_name: Mapped[str] = mapped_column(String, nullable=False)
-    amount: Mapped[float] = mapped_column(Numeric(12, 2), nullable=False)
+    # Integer MINOR units (cents for 2-decimal currencies, whole units for
+    # CLP/COP). Numeric returns Decimal at runtime — annotate honestly so type
+    # checkers flag float arithmetic on money.
+    amount: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
     currency: Mapped[str] = mapped_column(String, default="CLP")
     transaction_date: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False)
     category: Mapped[str | None] = mapped_column(String, nullable=True)  # denormalized
